added video streaming input

+++ utils/bbox_utils.py
@@ -16,7 +16,7 @@ def get_bbox_center(bbox):
         tuple: Center coordinates (x, y) of the bounding box.
     """
     x1,y1,x2,y2 = bbox
-    return int((x1+x2)/2), int((y1+y2)/2)
+    return ((x1+x2)/2), ((y1+y2)/2)
 
 def get_bbox_width(bbox):
     """
